+++ datafusion/src/expressions/mod.rs
@@ -1,8 +1,8 @@
-pub(crate) use self::delta_to_df::*;
-pub(crate) use self::df_to_delta::*;
+pub(crate) use self::to_datafusion::*;
+pub(crate) use self::to_delta::*;
 
-mod delta_to_df;
-mod df_to_delta;
+mod to_datafusion;
+mod to_delta;
 
 #[cfg(test)]
 #[cfg(test)]
@@ -15,7 +15,7 @@ mod tests {
     fn test_roundtrip_simple_and() {
         let df_expr = col("a").eq(lit(1)).and(col("b").eq(lit(2)));
         let delta_expr = to_delta_expression(&df_expr).unwrap();
-        let df_expr_roundtrip = to_df_expr(&delta_expr).unwrap();
+        let df_expr_roundtrip = to_datafusion_expr(&delta_expr).unwrap();
         assert_eq!(df_expr, df_expr_roundtrip);
     }
 
@@ -27,7 +27,7 @@ mod tests {
             .and(col("c").eq(lit(3)))
             .and(col("d").eq(lit(4)));
         let delta_expr = to_delta_expression(&df_expr).unwrap();
-        let df_expr_roundtrip = to_df_expr(&delta_expr).unwrap();
+        let df_expr_roundtrip = to_datafusion_expr(&delta_expr).unwrap();
         assert_eq!(df_expr, df_expr_roundtrip);
     }
 
@@ -38,7 +38,7 @@ mod tests {
             .and(col("b").eq(lit(2)))
             .or(col("c").eq(lit(3)).and(col("d").eq(lit(4))));
         let delta_expr = to_delta_expression(&df_expr).unwrap();
-        let df_expr_roundtrip = to_df_expr(&delta_expr).unwrap();
+        let df_expr_roundtrip = to_datafusion_expr(&delta_expr).unwrap();
         assert_eq!(df_expr, df_expr_roundtrip);
     }
 
@@ -46,7 +46,7 @@ mod tests {
     fn test_roundtrip_unary() {
         let df_expr = !col("a").eq(lit(1));
         let delta_expr = to_delta_expression(&df_expr).unwrap();
-        let df_expr_roundtrip = to_df_expr(&delta_expr).unwrap();
+        let df_expr_roundtrip = to_datafusion_expr(&delta_expr).unwrap();
         assert_eq!(df_expr, df_expr_roundtrip);
     }
 
@@ -54,7 +54,7 @@ mod tests {
     fn test_roundtrip_is_null() {
         let df_expr = col("a").is_null();
         let delta_expr = to_delta_expression(&df_expr).unwrap();
-        let df_expr_roundtrip = to_df_expr(&delta_expr).unwrap();
+        let df_expr_roundtrip = to_datafusion_expr(&delta_expr).unwrap();
         assert_eq!(df_expr, df_expr_roundtrip);
     }
 
@@ -62,7 +62,7 @@ mod tests {
     fn test_roundtrip_binary_ops() {
         let df_expr = col("a") + col("b") * col("c");
         let delta_expr = to_delta_expression(&df_expr).unwrap();
-        let df_expr_roundtrip = to_df_expr(&delta_expr).unwrap();
+        let df_expr_roundtrip = to_datafusion_expr(&delta_expr).unwrap();
         assert_eq!(df_expr, df_expr_roundtrip);
     }
 
@@ -70,7 +70,7 @@ mod tests {
     fn test_roundtrip_comparison_ops() {
         let df_expr = col("a").gt(col("b")).and(col("c").lt_eq(col("d")));
         let delta_expr = to_delta_expression(&df_expr).unwrap();
-        let df_expr_roundtrip = to_df_expr(&delta_expr).unwrap();
+        let df_expr_roundtrip = to_datafusion_expr(&delta_expr).unwrap();
         assert_eq!(df_expr, df_expr_roundtrip);
     }
 
@@ -82,21 +82,21 @@ mod tests {
             left: Box::new(Expression::Column("a".parse().unwrap())),
             right: Box::new(Expression::Column("b".parse().unwrap())),
         });
-        assert!(to_df_expr(&delta_expr).is_err());
+        assert!(to_datafusion_expr(&delta_expr).is_err());
 
         let delta_expr = Expression::Binary(BinaryExpression {
             op: BinaryOperator::In,
             left: Box::new(Expression::Column("a".parse().unwrap())),
             right: Box::new(Expression::Column("b".parse().unwrap())),
         });
-        assert!(to_df_expr(&delta_expr).is_err());
+        assert!(to_datafusion_expr(&delta_expr).is_err());
 
         let delta_expr = Expression::Binary(BinaryExpression {
             op: BinaryOperator::NotIn,
             left: Box::new(Expression::Column("a".parse().unwrap())),
             right: Box::new(Expression::Column("b".parse().unwrap())),
         });
-        assert!(to_df_expr(&delta_expr).is_err());
+        assert!(to_datafusion_expr(&delta_expr).is_err());
     }
 
     #[test]
@@ -105,6 +105,6 @@ mod tests {
             Expression::Column("a".parse().unwrap()),
             Expression::Column("b".parse().unwrap()),
         ]);
-        assert!(to_df_expr(&delta_expr).is_err());
+        assert!(to_datafusion_expr(&delta_expr).is_err());
     }
 }
